DEX-637 Backward compatibility tests for database
* The common part for backward compat tests of MultipleDifferentMatchersTestSuite is separated to BackwardCompatTestSuite;
* MultipleDifferentMatchersTestSuite is renamed to OrderBookBackwardCompatTestSuite;
